+++ pull_request_templete.md
@@ -23,9 +23,9 @@
 
 ## 🎯 확인해야 할 것
 
-- [ ] PR 내용과 실제 코드가 일치하는가?
-- [ ] 빌드/배포에 영향이 없는가?
-- [ ] 테스트가 필요한 부분은 모두 테스트했는가?
+- [x] PR 내용과 실제 코드가 일치하는가?
+- [x] 빌드/배포에 영향이 없는가?
+- [x] 테스트가 필요한 부분은 모두 테스트했는가?
 
 ---
